Simplify ItemThroughManager manager methods
Previously, the `exclude_*` methods did not handle the actual
negation of the matched records. As a result, this was left up to the
client to handle which is clearly a broken implementation. This has
been fixed and propagated through out the tests and translator. This
is technically a breaking change, but the methods have not been
considered public API so it is unlikely client applications will be
affected.