Refactor vocab resources to conform with Serrano's API
The ItemsResource is intended to replace the Serrano's FieldValuesResource
for vocab fields. The ItemResource exposes individual items
with links and other description data.

The separate search endpoint has been removed and is now implemented as
a GET parameter on the ItemsResource.

The arbitrary `max_results` has been removed now that the data is paginated
using Serrano's facilities.